LayoutTests/fast/mediastream/MediaStreamConstructor-expected.txt

Tests webkitMediaStream.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Got local stream.
PASS localStream.getAudioTracks().length is 1
PASS localStream.getVideoTracks().length is 1
PASS typeof webkitMediaStream === 'function' is true
PASS new webkitMediaStream(document) threw exception TypeError: Type error.
PASS new webkitMediaStream([document]) threw exception TypeError: Invalid Array element type.
PASS new webkitMediaStream([stream.getAudioTracks()[0], document]) threw exception TypeError: Invalid Array element type.
PASS new webkitMediaStream([null]) threw exception TypeError: Invalid Array element type.
PASS new webkitMediaStream([undefined]) threw exception TypeError: Invalid Array element type.
PASS new webkitMediaStream(null) threw exception TypeError: Type error.
PASS new webkitMediaStream(undefined) threw exception TypeError: Type error.
PASS Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS Stream constructed
PASS [object MediaStream] is non-null.
PASS [object MediaStream] is defined.
PASS newStream.constructor.name is 'MediaStream'
PASS newStream.getAudioTracks().length is nAudio
PASS newStream.getVideoTracks().length is nVideo
PASS successfullyParsed is true

TEST COMPLETE
